Moves token manager to cloud accounts
